fix(plugins): share route id normalization
