+++ src/plugins/activation-planner.ts
@@ -1,10 +1,15 @@
 import { normalizeProviderId } from "../agents/provider-id.js";
 import type { OpenClawConfig } from "../config/types.js";
 import { normalizeOptionalLowercaseString } from "../shared/string-coerce.js";
-import { loadPluginManifestRegistry, type PluginManifestRecord } from "./manifest-registry.js";
+import {
+  loadPluginManifestRegistry,
+  type PluginManifestRecord,
+  type PluginManifestRegistry,
+} from "./manifest-registry.js";
 import type { PluginManifestActivationCapability } from "./manifest.js";
 import type { PluginOrigin } from "./plugin-origin.types.js";
 import { createPluginIdScopeSet, normalizePluginIdScope } from "./plugin-scope.js";
+import { normalizeActivationRouteId } from "./route-id-normalize.js";
 
 export type PluginActivationPlannerTrigger =
   | { kind: "command"; command: string }
@@ -21,18 +26,22 @@ export function resolveManifestActivationPluginIds(params: {
   cache?: boolean;
   origin?: PluginOrigin;
   onlyPluginIds?: readonly string[];
+  manifestRegistry?: PluginManifestRegistry;
 }): string[] {
   const onlyPluginIdSet = createPluginIdScopeSet(normalizePluginIdScope(params.onlyPluginIds));
+  const manifestRegistry =
+    params.manifestRegistry ??
+    loadPluginManifestRegistry({
+      config: params.config,
+      workspaceDir: params.workspaceDir,
+      env: params.env,
+      cache: params.cache,
+    });
 
   return [
     ...new Set(
-      loadPluginManifestRegistry({
-        config: params.config,
-        workspaceDir: params.workspaceDir,
-        env: params.env,
-        cache: params.cache,
-      })
-        .plugins.filter(
+      manifestRegistry.plugins
+        .filter(
           (plugin) =>
             (!params.origin || plugin.origin === params.origin) &&
             (!onlyPluginIdSet || onlyPluginIdSet.has(plugin.id)) &&
@@ -118,12 +127,5 @@ function normalizeCommandId(value: string | undefined): string {
 }
 
 function normalizeRouteId(value: string | undefined): string {
-  const normalized = normalizeOptionalLowercaseString(value) ?? "";
-  switch (normalized) {
-    case "webhook":
-    case "gateway-webhook":
-      return "gateway-plugin-http";
-    default:
-      return normalized;
-  }
+  return normalizeActivationRouteId(value);
 }
