+++ src/plugins/manifest.ts
@@ -13,6 +13,7 @@ import {
 } from "./manifest-command-aliases.js";
 import type { PluginConfigUiHint } from "./manifest-types.js";
 import type { PluginKind } from "./plugin-kind.types.js";
+import { normalizeActivationRouteId } from "./route-id-normalize.js";
 
 export const PLUGIN_MANIFEST_FILENAME = "openclaw.plugin.json";
 export const PLUGIN_MANIFEST_FILENAMES = [PLUGIN_MANIFEST_FILENAME] as const;
@@ -432,7 +433,7 @@ function normalizeManifestActivation(value: unknown): PluginManifestActivation |
   const onRoutes = Array.from(
     new Set(
       normalizeTrimmedStringList(value.onRoutes).map((routeId) =>
-        normalizeManifestActivationRouteId(routeId),
+        normalizeActivationRouteId(routeId),
       ),
     ),
   );
@@ -454,17 +455,6 @@ function normalizeManifestActivation(value: unknown): PluginManifestActivation |
 
   return Object.keys(activation).length > 0 ? activation : undefined;
 }
-
-function normalizeManifestActivationRouteId(value: string): string {
-  switch (value.trim().toLowerCase()) {
-    case "webhook":
-    case "gateway-webhook":
-      return "gateway-plugin-http";
-    default:
-      return value;
-  }
-}
-
 function normalizeManifestSetupProviders(
   value: unknown,
 ): PluginManifestSetupProvider[] | undefined {
